test: add service tests

+++ src/test/java/it/polito/ezgas/UserConverterTest.java
@@ -7,6 +7,7 @@ import it.polito.ezgas.entity.User;
 import org.junit.Test;
 import static org.junit.jupiter.api.Assertions.*;
 
+
 public class UserConverterTest {
     @Test
     public void testAdminConvertToDto() {
@@ -15,11 +16,11 @@ public class UserConverterTest {
         user.setAdmin(true);
         user.setUserId(1);
         UserDto userDto = converter.convertToDto(user);
-        assertEquals(1, userDto.getUserId(), "admin dto userdId is wrong");
+        assertEquals((Integer)1, userDto.getUserId(), "admin dto userdId is wrong");
         assertEquals("admin", userDto.getUserName(), "admin dto username is wrong");
         assertEquals("adminpw", userDto.getPassword(), "admin dto password is wrong");
         assertEquals("admin@ezgas.com", userDto.getEmail(), "admin dto email is wrong");
-        assertEquals(5, userDto.getReputation(), "admin dto reputation is wrong");
+        assertEquals((Integer)5, userDto.getReputation(), "admin dto reputation is wrong");
         assertTrue(userDto.getAdmin(), "admin dto admin is wrong");
     }
 
@@ -32,7 +33,7 @@ public class UserConverterTest {
         assertEquals("admin", user.getUserName(), "admin username is wrong");
         assertEquals("adminpw", user.getPassword(), "admin password is wrong");
         assertEquals("admin@ezgas.com", user.getEmail(), "admin email is wrong");
-        assertEquals(5, user.getReputation(), "admin reputation is wrong");
+        assertEquals((Integer)5, user.getReputation(), "admin reputation is wrong");
         assertTrue(user.getAdmin(), "admin admin is wrong");
     }
 
@@ -44,7 +45,7 @@ public class UserConverterTest {
         assertEquals("admin", user.getUserName(), "admin username is wrong");
         assertEquals("adminpw", user.getPassword(), "admin password is wrong");
         assertEquals("admin@ezgas.com", user.getEmail(), "admin email is wrong");
-        assertEquals(5, user.getReputation(), "admin reputation is wrong");
+        assertEquals((Integer)5, user.getReputation(), "admin reputation is wrong");
         assertTrue(user.getAdmin(), "admin admin is wrong");
     }
 
@@ -55,11 +56,11 @@ public class UserConverterTest {
         user.setAdmin(false);
         user.setUserId(10);
         UserDto userDto = converter.convertToDto(user);
-        assertEquals(10, userDto.getUserId(), "user dto userdId is wrong");
+        assertEquals((Integer)10, userDto.getUserId(), "user dto userdId is wrong");
         assertEquals("deadpool", userDto.getUserName(), "user dto username is wrong");
         assertEquals("diehard", userDto.getPassword(), "user dto password is wrong");
         assertEquals("deadpool@marvel.com", userDto.getEmail(), "user dto email is wrong");
-        assertEquals(-3, userDto.getReputation(), "user dto reputation is wrong");
+        assertEquals((Integer)(-3), userDto.getReputation(), "user dto reputation is wrong");
         assertFalse(userDto.getAdmin(), "user dto admin is wrong");
     }
 
@@ -72,7 +73,7 @@ public class UserConverterTest {
         assertEquals("Alter Bridge", user.getUserName(), "user username is wrong");
         assertEquals("metalingus", user.getPassword(), "user password is wrong");
         assertEquals("ab@rock.com", user.getEmail(), "user email is wrong");
-        assertEquals(4, user.getReputation(), "user reputation is wrong");
+        assertEquals((Integer)4, user.getReputation(), "user reputation is wrong");
         assertFalse(user.getAdmin(), "user admin is wrong");
     }
 
@@ -84,7 +85,7 @@ public class UserConverterTest {
         assertEquals("Jebedaiah Kerman", user.getUserName(), "user username is wrong");
         assertEquals("MunCrash", user.getPassword(), "user password is wrong");
         assertEquals("jeb@ksp.com", user.getEmail(), "user email is wrong");
-        assertEquals(0, user.getReputation(), "user reputation is wrong");
+        assertEquals((Integer)0, user.getReputation(), "user reputation is wrong");
         assertFalse(user.getAdmin(), "user admin is wrong");
     }
 }
